core/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
billiard==4.2.1
boto3==1.37.32
botocore==1.37.32
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.1
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.9
Django==5.1.8
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-rest-knox==5.0.2
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
eval_type_backport==0.2.2
executing==2.2.0
filelock==3.18.0
gevent==25.5.1
gitdb==4.0.12
greenlet==3.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
identify==2.6.9
idna==3.10
inflection==0.5.1
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.2
matplotlib-inline==0.1.7
nodeenv==1.9.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
playwright==1.52.0
pre_commit==4.2.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
redis==6.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
ruff==0.11.5
s3transfer==0.11.4
setuptools==80.4.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
sseclient-py==1.8.0
stack-data==0.6.3
svix==1.64.1
traitlets==5.14.3
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2
